Renamed shuffle trait.
